server/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
django-rosetta==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
joblib==1.3.2
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
polib==1.2.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
